Update issue.updated_at if we have just marked an issue stale
This is to ensure we do not close issues before days-before-close has
expired. Without this fix we can encounter a situation where an issue
gets marked as stale and it gets closed immediately without waiting
days-before-close number of days.

If the user still wishes to close the issue immediately after marking
it stale they have the option of doing so by setting days-before-close
to 0 to get the desired effect.