Revest PR code suggestion
Reverted a code suggestion made in PR #4 to change github.context.actor to octomock.context.actor. This needed to be reverted to keep the github object with the correct context in the test as this represents the action's context and is part of what we are testing.
Changed the github import to be * as github to satisfy the needs of tests and provide the mocked action context.

Signed-off-by: Froilan Irizarry Rivera <1918027+froi@users.noreply.github.com>