Stop filtering comments by github.context.actor in determining activity
I added a scheduled workflow with the stale action and it seems that
when the action runs, `github.context.actor` is myself, not the github
actions bot. As a result, an issue that I had commented on after it was
labeled `stale` ended up getting closed anyway. Here's what's in the
logs for the action:
```
Checking for comments on issue #15419 since 2020-09-22T00:06:03Z
Comments not made by rmacklin or another bot: 0
```

According to this post in the GitHub Support Community, it seems to be
expected behavior that the actor will be the user who set up the
schedule:
https://github.community/t/who-will-be-the-github-actor-when-a-workflow-runs-on-a-schedule/17369

Thus, it doesn't seem appropriate to ignore comments from
`github.context.actor` when determining if there has been activity
since the issue was labeled stale; that is not a bot user as previously
thought when this behavior was originally added in
96b682d29f2b26168a8b283c62da1ad3f1262120.