feat: add support for end date
Add support for the end-date option. This option
is similar to the start-date option except is allows
skipping items which were created after the end date.

In the Node.js project we want to mark PRs which
are older than 1 year old and have not been updated in
6 months as stale and then close them. The end-date
option is needed to be able to only operate on PRs
which are older than one year.

Signed-off-by: Michael Dawson <mdawson@devrus.com>