Fix logging for removing nonexistent label
https://github.com/jsoref/stale-bot-debug/actions/runs/5828506506/job/15806335430#step:2:97

::group::[#2] Issue #2
[#2] Issue #2
...
  [#2] Removing all the labels specified via the labels-to-remove-when-stale (​https://github.com/actions/stale#labels-to-remove-when-stale​) option.
  [#2] Removing the label "label-to-add-when-unstale" from this issue...
  Error: [#2] Error when removing the label: "Label does not exist"

The quotes on this last line make it look like the action was trying to
remove a label named `Label does not exist`.

This change removes the quotes to make it clear that the text is just an
error message (which is implied by the `:`) as opposed to the name of a
label, which is the reading one gets from the inclusion of additional
quotes around the text.